rename and move verions.go to greenplum package
This allows access by both the hub and agent.

Rename greenplum.ValidateVersions to
greenplum.VerifyCompatibleGPDBVersions for clarity. This is to
distinguish from validating gpupgrade and GPDB versions across hosts.
